backend/pyproject.toml
[project]
name = "joker_task"
version = "0.6.0"
description = ""
authors = [
    {name = "bruno-gabriel-muniz",email = "brunomunizlourenco@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13, <4.0"
dependencies = [
    "fastapi[standard] (>=0.117.1,<0.118.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pwdlib[argon2] (>=0.2.1,<0.3.0)",
    "sqlalchemy[asyncio] (>=2.0.43,<3.0.0)",
    "alembic (>=1.16.5,<2.0.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
pytest-cov = "^7.0.0"
taskipy = "^1.14.1"
ruff = "^0.13.2"
pytest-asyncio = "^1.2.0"
ipdb = "^0.13.13"
freezegun = "^1.5.5"
testcontainers = "^4.14.1"
asyncpg = {extras = ["gssauth"], version = "^0.31.0"}
psycopg = "^3.2.10"

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'
asyncio_default_fixture_loop_scope = 'function'

[tool.coverage.run]
concurrency = ["thread", "greenlet"]

[tool.taskipy.tasks]
lint = 'ruff check'
format = 'ruff format'
post_format = 'ruff check --fix'
run = 'fastapi dev backend/joker_task/app.py'
pre_testf = 'task format'
testf = 'pytest -s -x --cov=joker_task -vv'
post_testf = 'coverage html'

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
